fix test so file gets tested
